backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
bottle==0.12.25
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
fasttext-wheel==0.9.2
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mcpi==1.2.1
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nltk==3.8.1
node-semver==0.6.1
numpy==1.24.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
patch-ng==1.17.4
pluginbase==1.0.1
protobuf==4.25.3
pybind11==2.12.0
Pygments==2.18.0
pywin32==306
regex==2024.5.10
requests==2.28.2
rich==13.7.1
sanic-routing==0.7.2
scipy==1.9.3
shutup==0.2.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==1.26.14
wcwidth==0.2.12
Werkzeug==3.0.3
wrapt==1.16.0